Fixed attachment buttons non-IE Browsers
On the new post item for Forums, for non-IE browsers the text of the file name didn't show.  Also, the alignment was wonky.  

Modified line 233 to remove out the reference to the class in the CSS.  This was causing the button to be centered but depending on the browser the file name would be left or right and would look out of place.
Modified line 234 to remove "textbox" class so it will display correct in Non-IE browser.
Modified line 303 for the same reason as 234.